Consolidate dashboard and scheduling into send page
