+++ src/admin/src/App.tsx
@@ -1,22 +1,12 @@
 import React from "react";
 import { NavLink, Routes, Route } from "react-router-dom";
-import {
-  HomeIcon,
-  PaperAirplaneIcon,
-  ClockIcon,
-  BellIcon,
-  QrCodeIcon,
-} from "./lib/icons";
-import Dashboard from "./pages/Dashboard";
+import { PaperAirplaneIcon, BellIcon, QrCodeIcon } from "./lib/icons";
 import Send from "./pages/Send";
-import Scheduling from "./pages/Scheduling";
 import Alerts from "./pages/Alerts";
 import Qr from "./pages/Qr";
 
 const navItems = [
-  { to: "/", label: "Dashboard", icon: HomeIcon },
-  { to: "/send", label: "Send", icon: PaperAirplaneIcon },
-  { to: "/scheduling", label: "Scheduling", icon: ClockIcon },
+  { to: "/", label: "Send", icon: PaperAirplaneIcon },
   { to: "/alerts", label: "Alerts", icon: BellIcon },
   { to: "/qr", label: "QR", icon: QrCodeIcon },
 ];
@@ -52,9 +42,8 @@ export default function App() {
       </nav>
       <main className="flex-1 p-6 overflow-auto">
         <Routes>
-          <Route path="/" element={<Dashboard />} />
+          <Route path="/" element={<Send />} />
           <Route path="/send" element={<Send />} />
-          <Route path="/scheduling" element={<Scheduling />} />
           <Route path="/alerts" element={<Alerts />} />
           <Route path="/qr" element={<Qr />} />
         </Routes>
