+++ src/admin/src/pages/ScheduleList.tsx
@@ -1,16 +1,12 @@
 import React, { useEffect, useState } from "react";
 import {
   listSchedules,
-  createSchedule,
   deleteSchedule,
   runSchedule,
   pauseSchedule,
   resumeSchedule,
   updateSchedule,
-  fetchTopContacts,
-  fetchAllContacts,
 } from "../lib/api";
-import type { Contact } from "../lib/types";
 
 interface Schedule {
   id: string;
@@ -25,21 +21,12 @@ interface Schedule {
   lastRunAt: string | null;
 }
 
-export default function Scheduling() {
+interface Props {
+  refreshKey: number;
+}
+
+export default function ScheduleList({ refreshKey }: Props) {
   const [schedules, setSchedules] = useState<Schedule[]>([]);
-  const [form, setForm] = useState({
-    phone: "",
-    text: "",
-    enablePrefix: false,
-    firstRunAt: "",
-    intervalValue: "" as string | number,
-    intervalUnit: "minutes" as "minutes" | "hours" | "days",
-    active: true,
-  });
-  const [selected, setSelected] = useState<Contact | null>(null);
-  const [topContacts, setTopContacts] = useState<Contact[]>([]);
-  const [allContacts, setAllContacts] = useState<Contact[]>([]);
-  const [showAll, setShowAll] = useState(false);
   const [edit, setEdit] = useState<{
     id: string;
     text: string;
@@ -49,25 +36,10 @@ export default function Scheduling() {
     intervalUnit: "minutes" | "hours" | "days";
     active: boolean;
   } | null>(null);
-  const [status, setStatus] = useState<string | null>(null);
-
-  const suggestions = selected
-    ? []
-    : form.phone
-      ? allContacts.filter(
-          (c) =>
-            c.phone?.includes(form.phone) ||
-            c.name.toLowerCase().includes(form.phone.toLowerCase()),
-        )
-      : showAll
-        ? allContacts
-        : topContacts;
 
   useEffect(() => {
     refresh();
-    fetchTopContacts().then((res) => setTopContacts(res.contacts));
-    fetchAllContacts().then((res) => setAllContacts(res.contacts));
-  }, []);
+  }, [refreshKey]);
 
   async function refresh() {
     try {
@@ -75,51 +47,6 @@ export default function Scheduling() {
       setSchedules(data.items);
     } catch (err) {
       console.error(err);
-    }
-  }
-
-  async function handleCreate(e: React.FormEvent) {
-    e.preventDefault();
-    try {
-      const payload: any = {
-        text: form.text,
-        enablePrefix: form.enablePrefix,
-        active: form.active,
-      };
-      if (selected) {
-        if (selected.phone) payload.phone = selected.phone;
-        else payload.name = selected.name;
-      } else if (form.phone) {
-        payload.phone = form.phone;
-      } else {
-        throw new Error("Missing contact");
-      }
-      if (form.firstRunAt)
-        payload.firstRunAt = new Date(form.firstRunAt).toISOString();
-      if (form.intervalValue) {
-        const mult =
-          form.intervalUnit === "hours"
-            ? 60
-            : form.intervalUnit === "days"
-              ? 1440
-              : 1;
-        payload.intervalMinutes = Number(form.intervalValue) * mult;
-      }
-      await createSchedule(payload);
-      setStatus("Schedule created");
-      setForm({
-        phone: "",
-        text: "",
-        enablePrefix: false,
-        firstRunAt: "",
-        intervalValue: "",
-        intervalUnit: "minutes",
-        active: true,
-      });
-      setSelected(null);
-      refresh();
-    } catch (err: any) {
-      setStatus(err.response?.data?.error || err.message || "Error creating");
     }
   }
 
@@ -172,136 +99,6 @@ export default function Scheduling() {
 
   return (
     <div className="space-y-6">
-      <h1 className="text-2xl font-bold">Scheduling</h1>
-      {/* Create form */}
-      <div className="bg-wa-panel p-4 rounded-lg space-y-4 max-w-xl">
-        <h2 className="text-lg font-medium">Create Schedule</h2>
-        <form onSubmit={handleCreate} className="space-y-3">
-          <div>
-            <label className="block mb-1">Contact</label>
-            <input
-              type="text"
-              value={form.phone}
-              onChange={(e) => {
-                setForm({ ...form, phone: e.target.value });
-                setSelected(null);
-                if (e.target.value) setShowAll(false);
-              }}
-              className="w-full px-3 py-2 rounded bg-wa-hover text-white"
-            />
-            {suggestions.length > 0 ? (
-              <div className="mt-2 max-h-48 overflow-y-auto space-y-2">
-                {suggestions.map((c) => (
-                  <button
-                    key={c.phone || c.name}
-                    type="button"
-                    onClick={() => {
-                      setSelected(c);
-                      setForm({
-                        ...form,
-                        phone: c.phone ? `${c.name} (${c.phone})` : c.name,
-                      });
-                    }}
-                    className="block w-full text-left bg-wa-hover hover:bg-wa-panel p-2 rounded"
-                  >
-                    <div className="font-medium">{c.name}</div>
-                    {c.phone && (
-                      <div className="text-xs text-gray-400">{c.phone}</div>
-                    )}
-                  </button>
-                ))}
-              </div>
-            ) : (
-              !form.phone && (
-                <p className="text-sm text-gray-400 mt-2">
-                  No contacts yet. Add or import contacts to get started.
-                </p>
-              )
-            )}
-            {!form.phone &&
-              !showAll &&
-              allContacts.length > topContacts.length && (
-                <button
-                  type="button"
-                  onClick={() => setShowAll(true)}
-                  className="text-wa-green text-sm mt-2"
-                >
-                  View all
-                </button>
-              )}
-          </div>
-          <div>
-            <label className="block mb-1">Message</label>
-            <textarea
-              value={form.text}
-              onChange={(e) => setForm({ ...form, text: e.target.value })}
-              className="w-full px-3 py-2 rounded bg-wa-hover text-white"
-            />
-          </div>
-          <div className="flex items-center space-x-2">
-            <input
-              type="checkbox"
-              checked={form.enablePrefix}
-              id="createEnablePrefix"
-              onChange={(e) =>
-                setForm({ ...form, enablePrefix: e.target.checked })
-              }
-            />
-            <label htmlFor="createEnablePrefix">Enable prefix</label>
-          </div>
-          <div>
-            <label className="block mb-1">First run (local)</label>
-            <input
-              type="datetime-local"
-              value={form.firstRunAt}
-              onChange={(e) => setForm({ ...form, firstRunAt: e.target.value })}
-              className="w-full px-3 py-2 rounded bg-wa-hover text-white"
-            />
-          </div>
-          <div>
-            <label className="block mb-1">Interval</label>
-            <div className="flex space-x-2">
-              <input
-                type="number"
-                min="0"
-                value={form.intervalValue}
-                onChange={(e) =>
-                  setForm({ ...form, intervalValue: e.target.value })
-                }
-                className="w-full px-3 py-2 rounded bg-wa-hover text-white"
-              />
-              <select
-                value={form.intervalUnit}
-                onChange={(e) =>
-                  setForm({ ...form, intervalUnit: e.target.value as any })
-                }
-                className="px-2 py-2 rounded bg-wa-hover text-white"
-              >
-                <option value="minutes">Minutes</option>
-                <option value="hours">Hours</option>
-                <option value="days">Days</option>
-              </select>
-            </div>
-          </div>
-          <div className="flex items-center space-x-2">
-            <input
-              type="checkbox"
-              checked={form.active}
-              id="createActive"
-              onChange={(e) => setForm({ ...form, active: e.target.checked })}
-            />
-            <label htmlFor="createActive">Active</label>
-          </div>
-          <button
-            type="submit"
-            className="bg-wa-green hover:bg-wa-green/80 px-4 py-2 rounded text-wa-bg"
-          >
-            Create
-          </button>
-          {status && <p className="text-yellow-400 text-sm">{status}</p>}
-        </form>
-      </div>
-      {/* List */}
       <div className="bg-wa-panel p-4 rounded-lg">
         <h2 className="text-lg font-medium mb-4">Existing Schedules</h2>
         <div className="overflow-x-auto">
@@ -401,7 +198,6 @@ export default function Scheduling() {
           </table>
         </div>
       </div>
-      {/* Edit modal */}
       {edit && (
         <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
           <div className="bg-wa-panel p-6 rounded-lg space-y-4 w-96">
